Move ParallelCollection into spark.rdd package.
